requirements.txt
aiogram
google.generativeai
python-dotenv
asyncpg
Pillow